Bring up to rev 27582
git-svn-id: https://svn.cv.nrao.edu/svn/casa/branches/project@27583 ac9c0da5-131f-0410-9009-a9c195f183e6

+++ darwin13port/code/air_casawvr/cmdline/wvrgcal.cpp
@@ -284,7 +284,7 @@ void flagInterp(const casa::MeasurementSet &ms,
 						wvrflag_s, 
 						3,
 						maxdist_m);
-    if(near.size()>=minnumants){
+    if(near.size()>= static_cast<unsigned int>(minnumants)){
       LibAIR::interpBadAntW(d, 
 			    *i, 
 			    near);
